+++ src/main/resources/application.yml
@@ -101,8 +101,6 @@ cache:
   expire-after-write: 6h
 
 management:
-  server:
-    port: ${BACKEND_ACTUATOR_PORT:8078}
   endpoints:
     web:
       exposure:
